Added template parameter to csa_wt which allows to
represent the sentinel implicitly. I.e. the sentinel
is not stored explicitly in the BWT. It is implicitly
represented by storing its position in the BWT.
This reduces the effective alphabet size by one.
However rank/select/access operations might be slower,
since we have to check if the requested information
is left or right of the position of the sentinel.